feat: implement PDF export functionality with inline text extraction, image handling, and rendering
- Add inline text extraction logic to convert DOM elements into styled text fragments.
- Implement image acquisition helpers to preload and convert images to data URLs.
- Create a renderer class to handle page-by-page rendering of blocks, including headings, paragraphs, lists, code, tables, admonitions, and images.
- Define shared types for the PDF generation pipeline, including inline and block models.
- Introduce utility functions for filename sanitization, text manipulation, link normalization, and color handling.
